Adding MetaFile class
